feat(retry): migrate to github.com/sethvargo/go-retry (#1139)
https://github.com/sethvargo/go-retry

+++ retry/config.go
@@ -4,5 +4,5 @@ package retry
 type Config struct {
 	Timeout  string `yaml:"timeout,omitempty" json:"timeout,omitempty" toml:"timeout,omitempty"`
 	Backoff  string `yaml:"backoff,omitempty" json:"backoff,omitempty" toml:"backoff,omitempty"`
-	Attempts uint   `yaml:"attempts,omitempty" json:"attempts,omitempty" toml:"attempts,omitempty"`
+	Attempts uint64 `yaml:"attempts,omitempty" json:"attempts,omitempty" toml:"attempts,omitempty"`
 }
